[ios][media-library] create utils file
